♻️ refactor : Category 관련 변수 ID화 - docker-compose내 version 삭제
