+++ docker/docker-compose.yml
@@ -1,5 +1,3 @@
-version: '3.1'
-
 services:
   mysql:
     image: mysql:latest
